Isolate LDAP state and split slow specs
